Add PHP version as optional action input
Psalm allows us to specify the PHP version it analyzes code against.
Exposing this option allows users of this action to supply a matrix of
PHP options to enable Psalm runs against each of the PHP versions they
would like to support, or just to analyze against a specific version of
PHP that differs from the container's installed version.